fix: Suggested fixes and additional testing round arm build

+++ src/llrt-function.ts
@@ -1,5 +1,5 @@
 import { posix } from 'path';
-import { CfnResource, Stack } from 'aws-cdk-lib';
+import { CfnResource, Stack, ValidationError } from 'aws-cdk-lib';
 import { Architecture, Code, ILayerVersion, LayerVersion, Runtime, RuntimeFamily } from 'aws-cdk-lib/aws-lambda';
 import { ICommandHooks, NodejsFunction, NodejsFunctionProps, OutputFormat } from 'aws-cdk-lib/aws-lambda-nodejs';
 import { Construct } from 'constructs';
@@ -49,7 +49,9 @@ export interface LlrtFunctionProps extends NodejsFunctionProps {
 
   /**
    * If `true` then the LLRT runtime will be built in a layer that can be shared amongst
-   * other `LLrtFunction`s that utilise the same `LlrtBinaryType` (this property will be ignored if `llrtBinaryPath` is set).
+   * other `LLrtFunction`s that utilise the same `llrtBinaryType`, `llrtVersion` and `architecture`.
+   * 
+   * This feature cannot be used with `llrtBinaryPath` and if both are set a ValidationError will be thrown.
    * 
    * @default - false
    */
@@ -59,7 +61,7 @@ export interface LlrtFunctionProps extends NodejsFunctionProps {
 export class LlrtFunction extends NodejsFunction {
   constructor(scope: Construct, id: string, props: LlrtFunctionProps) {
     const version = props.llrtVersion ?? 'latest';
-    const arch = props.architecture == Architecture.ARM_64 ? 'arm64' : 'x64';
+    const arch = props.architecture?.name == Architecture.ARM_64.name ? 'arm64' : 'x64';
     const binaryType = props.llrtBinaryType ?? LlrtBinaryType.STANDARD;
 
     let binaryName: string;
@@ -125,19 +127,26 @@ export class LlrtFunction extends NodejsFunction {
     const cacheDir = `.tmp/llrt/${version}/${arch}/${binaryType}`;
 
     const { commandHooks: originalCommandHooks, ...otherBundlingProps } = props.bundling ?? {};
-    const afterBundlingCommandHook: ICommandHooks['afterBundling'] = (i, o) => props.llrtBinaryPath ? [`cp ${posix.join(i, props.llrtBinaryPath)} ${posix.join(o, 'bootstrap')}`] :
-     props.useLambdaLayer ? [] : [
+    const afterBundlingCommandHook: ICommandHooks['afterBundling'] = (i, o) => {
+      // If useLambdaLayer then we'll package the llrt binary with docker later
+      if (props.useLambdaLayer) return [];
+      // Copy llrt binary from llrtBinaryPath
+      if (props.llrtBinaryPath) return [
+        `cp ${posix.join(i, props.llrtBinaryPath)} ${posix.join(o, 'bootstrap')}`,
+      ];
       // Download llrt binary from GitHub release and cache it
-      `if [ ! -e ${posix.join(i, cacheDir, 'bootstrap')} ]; then
-        mkdir -p ${posix.join(i, cacheDir)}
-        cd ${posix.join(i, cacheDir)}
-        curl -L -o llrt_temp.zip ${binaryUrl}
-        unzip llrt_temp.zip
-        rm -rf llrt_temp.zip
-        cd -
-       fi`,
-      `cp ${posix.join(i, cacheDir, 'bootstrap')} ${o}`,
-    ];
+      return [
+        `if [ ! -e ${posix.join(i, cacheDir, 'bootstrap')} ]; then
+          mkdir -p ${posix.join(i, cacheDir)}
+          cd ${posix.join(i, cacheDir)}
+          curl -L -o llrt_temp.zip ${binaryUrl}
+          unzip llrt_temp.zip
+          rm -rf llrt_temp.zip
+          cd -
+        fi`,
+        `cp ${posix.join(i, cacheDir, 'bootstrap')} ${o}`,
+      ];
+    }
 
     super(scope, id, {
       // set this to remove an unnecessary environment variable.
@@ -163,7 +172,11 @@ export class LlrtFunction extends NodejsFunction {
       },
     });
 
-    if (props.useLambdaLayer && !props.llrtBinaryPath) {
+    if (props.useLambdaLayer && props.llrtBinaryPath) {
+      throw new ValidationError("useLambdaLayer not supported with llrtBinaryPath", this);
+    }
+
+    if (props.useLambdaLayer) {
       this.ensureLayer(binaryUrl, binaryName, version);
     }
 
@@ -173,7 +186,7 @@ export class LlrtFunction extends NodejsFunction {
   private ensureLayer(binaryUrl: string, binaryName: string, version: string) {
     const id = `${binaryName.replace("lambda", "layer")}-${version}`;
     
-    let layer = Stack.of(this).node.tryFindChild(id);
+    let layer = Stack.of(this).node.tryFindChild(id) as ILayerVersion;
 
     if (!layer) {
       layer = new LayerVersion(Stack.of(this), id, {
@@ -181,11 +194,14 @@ export class LlrtFunction extends NodejsFunction {
           buildArgs: {
             URL: binaryUrl,
           },
+          file: "layer.Dockerfile",
         }),
-        layerVersionName: id,
+        compatibleArchitectures: [
+          this.architecture,
+        ],
       });
     }
 
-    this.addLayers(layer as ILayerVersion);
+    this.addLayers(layer);
   }
 }
